MDEV-31851   After crash recovery, undo tablespace fails to open
Problem:
========
- InnoDB fails to open undo tablespace when page0 is corrupted
and fails to throw error.

Solution:
=========
- InnoDB throws DB_CORRUPTION error when InnoDB encounters
page0 corruption of undo tablespace.

- InnoDB restores the page0 of undo tablespace from
doublewrite buffer if it encounters page corruption

- Moved Datafile::restore_from_doublewrite() to
recv_dblwr_t::restore_first_page(). So that undo
tablespace and system tablespace can use this function
instead of duplicating the code

srv_undo_tablespace_open(): Returns 0 if file doesn't exist
or ~0 if page0 is corrupted.

Changed datatype of the variables srv_undo_space_id_start,
srv_undo_tablespaces_open, srv_undo_tablespaces to uint32_t

+++ storage/innobase/include/srv0srv.h
@@ -250,7 +250,7 @@ extern bool	srv_was_started;
 extern char*	srv_undo_dir;
 
 /** Number of undo tablespaces to use. */
-extern ulong	srv_undo_tablespaces;
+extern uint	srv_undo_tablespaces;
 
 /** The number of UNDO tablespaces that are active (hosting some rollback
 segment). It is quite possible that some of the tablespaces doesn't host
